unversity adimission(web application)/project/requirements.txt
blinker==1.4
cassandra-driver==3.25.0
certifi==2020.12.5
click==8.0.1
Flask==2.0.1
Flask-Mail==0.9.1
geomet==0.2.1.post1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
six==1.16.0
Werkzeug==2.0.1
